DATAisFORMAT {MPAtools}	R Documentation

DATA is FORMAT

Description

Evaluates if a passed data.frame has a specified format

Usage

DATAisFORMAT(data, format)

Arguments

data
a data.frame that will be compared to a data.frame with the specified format as specified by the COBIApp

format
The specified format to which data will be compared

Value

message an error message displayed if the data is not consistent with format

Author(s)

Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>

[Package MPAtools version 0.1.0 Index]