b2c {MPAtools}	R Documentation

Convert data format

Description

Convert fish transect data from format b to format c; See details. Convierte datos de transectos de peces del formato b al formato c; Ver detalles

Usage

b2c(b)

Arguments

a
A data.frame with the format b. Un data.frame en formato b

Value

A data.frame wih the format c. Un data.frame en formato c

Author(s)

Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>

Examples

#library(MPAtools)

#load a "b" sample dataset
data("b")

#convert a to b format
c <- b2c(b)
c

[Package MPAtools version 0.1.0 Index]