bio_model {MPAtools}	R Documentation

Fit a generalized linear model

Description

This fits a linear model with the form of Indicador ~ Ano + Zona + Post * Zona + Temperatura + Visibilidad + Profundidad to the data. Depending on the specified covars, Temperature, Visibility and Depth can be dropped from the model. Use define_covars to identify which covariates to keep or specify them manually.

Usage

bio_model(data, covars = c("Temperatura", "Visibilidad", "Profundidad"))

Arguments

data
A data.frame containing at least the columns for Indicador, Ano, Zona, Post, Temperature, Visibility, and Profundidad. The indicator can be any of the biophysical indicators generated by shannon(), density(), fish_biomass(), trophic(), fish_size()

covars
a character vector specifying which covariates to keep. Options are Temperatura, Profundidad, Visibilidad or None (to drop all covariates).

Value

An object of class lm containing the fitted model

Author(s)

Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>

[Package MPAtools version 0.1.0 Index]