bio_results {MPAtools}	R Documentation

Biophysical Results

Description

Analyzes all the required biophysical indicators at once

Usage

bio_results(values, data, res, con)

Arguments

values
A list, created within the Shiny app, that contains the list of indicators selected, the community, any objective species, and the year of implementation of the reserve

data
A data.frame containing the raw data to be analyzed

res
A string specifying the name of the reserve site(s)

con
A string specifying the name of the control site(s)

Value

results, a tibble containing the name of the indicator as well as DiD estimates, p values, a string to be used as input for valueBox, the color of the valueBox, and the lm object fit to the indicator extracted from the data

Author(s)

Villasenor-Derbez, J.C. <jvillasenor@turfeffect.org>

[Package MPAtools version 0.1.0 Index]