c2b {MPAtools}	R Documentation

Convert data format

Description

Convert fish transect data from format c to format b; See details. Convierte datos de transectos de peces del formato c al formato b; Ver detalles

Usage

c2b(c)

Arguments

a
A data.frame with the format c. Un data.frame en formato c

Value

A data.frame wih the format b. Un data.frame en formato b

Author(s)

Villaseñor-Derbez, J.C. <jvillasenor@turfeffect.org>

Examples

#library(MPAtools)

#load an "c" sample dataset
data("c")

#convert c to a format
b <- c2b(c)
b

[Package MPAtools version 0.1.0 Index]