covars_wrapper {MPAtools}	R Documentation

A wrapper for define_covars

Description

Identifies for each sampling point (before-after-control-reserve) which covariates to use using define_covars. It then identifies the common covariates available for each sampling point and returns a string that bio_model can use to implement the correct model

Usage

covars_wrapper(data, year.imp)

Arguments

data
dataset

year.imp
year of implementation

Value

a vector stating which covariates to be used

[Package MPAtools version 0.1.0 Index]