density {MPAtools}	R Documentation

Calculate density

Description

Calculates species density (organisms / transect), by year, site, and transect, for a specific location.

Usage

density(data, location, species = NULL)

Arguments

data
A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie, Abundancia.

location
A quoted string that indicates the location.

species
A quoted string that indicates a species for which density should be calulated.

Value

D A dataframe with columns for Ano, Zona, Sitio, Transecto, GeneroEspecie, and D (density; org/m2).

Author(s)

Villaseñor-Derbez, J.C. <juancarlos.villader@gmail.com>

[Package MPAtools version 0.1.0 Index]