fish_biomass {MPAtools}	R Documentation

Calculate biomass

Description

Calulates biomass for a selected species or the entire dataset, by year, site, and transect, for a given location.

Usage

fish_biomass(data, location, species = NULL)

Arguments

data
A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie. Columns for a and b parameters are optional, though heavily suggested to avoid errors in calculations (this package might not have all parameters for all species).

location
A quoted string that indicates the location.

species
A quoted string that indicates a species for which density should be calulated.

a, b
An optional dataframe that contains the a and b allometric parameters for each species. Dataframe must have three columns: GeneroEspecie, a, b

Value

size A dataframe with columns for Ano, Zona, Sitio, Transecto, GeneroEspecie, and fish biomass (B; in grams).

[Package MPAtools version 0.1.0 Index]