fish_size {MPAtools}	R Documentation

Calculate a descriptor for fish size

Description

Calulates mean, median, maximum, or minimum size (cm) for a list of species (or selected species, as indicated by parameter species) by year, site, and transect, for a speciic location.

Usage

fish_size(data, location, species = NULL)

Arguments

data
A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie, Abundancia.

location
A quoted string that indicates the location.

species
A quoted string that indicates a species for which the size statistic should be calulated.

Value

size A dataframe with columns for Ano, Zona, Sitio, Transecto, GeneroEspecie, and the statistic required.

[Package MPAtools version 0.1.0 Index]