gen_score {MPAtools}	R Documentation

General Score

Description

According to the percentage of indicators that are positive, return a color based on a red - orange - yellow - green - olive scale, define by 20

Usage

gen_score(percent)

Arguments

percent
A number, between 0 and 100 that represents the percentage of positive indicators

Value

a color

[Package MPAtools version 0.1.0 Index]