mpa_area {MPAtools}	R Documentation

area matrix generator

Description

Generates the area matrix that will be simulated, based on size of the area, size of the MPA, number of MPAs,harvesting rate, and harvesting strategy.

Usage

mpa_area(m, n, mpa.size, mpa.number, harvesting.rate = 0.1,
  harvesting.strategy = FALSE)

Arguments

m
number of vertical cells (rows)

n
number of horizontal cells (columns)

mpa.size
proportion of cells to be MPAs (from 0 to 1)

mpa.number
number of MPAs

harvesting.rate
harvesting rate expressed as a proportion (0 to 1) of organisms harvested per cell. If harvesting.strategy = TRUE, this is the maximum value observed in the borders of the MPA.

harvesting.strategy
A logical value that indicates if "Fising the line" occurs. harvesting.strategy=FALSE is the default that indicates equal harvesting rate outside the MPA. If TRUE, it creates a pattern in which harvesting rate depends on the distance to MPA.

Value

area A matriz of size m by n that simulates harvesting rates and MPAs

Author(s)

Villasenor-Derbez, J.C.

[Package MPAtools version 0.1.0 Index]