mpa_check {MPAtools}	R Documentation

mpa_sim input validator

Description

Validates the inputs passed to mpa_2D().

Usage

mpa_check(area, nsteps, r, pop0, K, mrate)

Arguments

area
an N by M matrix with harvesting rates for fishing grounds (0 < harvesting rate <= 1) and MPAs (harvesting rate = 0) for the simulated area.

nsteps
Number of steps to run the simulation.

r
Intrinsic population growth rate.

pop0
Initial population size. If single, assumes equal population size across all cells in area. If a matrix, it must have the same dimensions as area.

K
Carrying capacity

mrate
Movement rate expressed as a proportion (relative to 1). Indicates the proportion of organisms that will move out of a cell in one timestep.

op
A logical value that indicates if a filled.contour is to be created. WARNING: op = TRUE may make the simulation slow.

Value

check an object of type list containing elemetns result and message. Result is a logical that indicates if the inputs to mpa_2d() are correct. If they are not, message contains the error message to be displayed in the Console.

Author(s)

Villasenor-Derbez, J.C.

[Package MPAtools version 0.1.0 Index]