mpa_plot3 {MPAtools}	R Documentation

Plots results from biophysical indicator functions by site

Description

Plots results from the functions used to calculate biophysical indicators like: richness(), density(), ttrophic(), fish_biomass(), and fish_size().

Usage

mpa_plot3(data, reserve = NULL, control = NULL, error.bars = F,
  y.lab = "Indicador")

Arguments

data
dataframe generated by any of the following functions: richness(), density(), trophic(), fish_biomass(), and fish_size().

reserve
A quoted string that indicates which site should be used as a reserve.

control
A quoted string that indicates which site should be used as a control zone for the indicated reserve.

error.bars
A logical value indicating if error bars (1 SD) should be plotted.

x.lab
A string to be used as x-axis label

Author(s)

Villasenor-Derbez, J.C.

[Package MPAtools version 0.1.0 Index]