mpa_plot4 {MPAtools}	R Documentation

Plot a turfeffect object for biophysical indicators

Description

Plot a turfeffect object for biophysical indicators

Usage

mpa_plot4(model, y.lab = NULL)

Arguments

model
An object of class lm, generated by the turfeffect() function or extracted from the tibble output of the bio_results() function

y.lab
A character that specifies the y.label

Value

an object og class ggplot2

[Package MPAtools version 0.1.0 Index]