richness {MPAtools}	R Documentation

Calculate species richness

Description

Calculates species richness, by year, site, and transect, for a specific location.

Usage

richness(data, location)

Arguments

data
A dataframe that contains at least columns of Ano, Zona (e.g. "Zona de pesca" or "Zona de no pesca"), Sitio, Transecto, GeneroEspecie.

location
A quoted string that indicates the location.

Value

S A dataframe with columns for Ano, Zona, Sitio, Transecto, and Species richness (S; number of species).

[Package MPAtools version 0.1.0 Index]