soc_plot {MPAtools}	R Documentation

Plot a turfeffect object for socioeconomic indicators

Description

Plot a turfeffect object for socioeconomic indicators

Usage

soc_plot(data, y.lab = "Value")

Arguments

y.lab
A character that specifies the y label

model
An object of class lm, generated by the turfeffect() function or extracted from the tibble output of the bio_results() function

Value

an object of class ggplot2

[Package MPAtools version 0.1.0 Index]