soc_results {MPAtools}	R Documentation

Results

Description

Analyzes all the required biophysical indicators at once

Usage

soc_results(values, data)

Arguments

values
A list, created within the Shiny app, that contains the list of indicators selected and the community

data
The data to be analyzed

Value

results, a tibble containing the name of the indicator as well as DiD estimates, p values, a string to be used as input for valueBox, the color of the valueBox, and the lm object fit to the indicator extracted from the data

[Package MPAtools version 0.1.0 Index]