valueBoxString {MPAtools}	R Documentation

valueBoxValues

Description

Calculate color and string needed in the ShinyApp valueBoxOutput based on a turfeffect model output

Usage

valueBoxString(model, type = NULL)

Arguments

model
An object of class lm, generated by the turfeffect function

type
A string

Value

x, a list with the first element (x$x) containing the string of slope and p value, and a second element (x$color) containing the color that the valueBox should have

[Package MPAtools version 0.1.0 Index]